requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
telethon==1.33.0
aiohttp==3.9.0
cryptography==41.0.7
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
